+++ js/modules/more.js
@@ -2,6 +2,8 @@
 const More = (() => {
   const ITEMS = () => [
     { id: 'global', t: I18n.t('tools.global.t'), d: I18n.t('tools.global.d') },
+    { id: 'commodities', t: I18n.t('tools.commodities.t'), d: I18n.t('tools.commodities.d') },
+    { id: 'announcements', t: I18n.t('tools.announcements.t'), d: I18n.t('tools.announcements.d') },
     { id: 'zakat', t: I18n.t('tools.zakat.t'), d: I18n.t('tools.zakat.d') },
     { id: 'import', t: I18n.t('tools.import.t'), d: I18n.t('tools.import.d') },
     { id: 'screener', t: I18n.t('tools.screener.t'), d: I18n.t('tools.screener.d') },